+++ lib/src/upgrader.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:package_info/package_info.dart';
@@ -18,6 +19,9 @@ import 'upgrade_messages.dart';
 
 /// Signature of callbacks that have no arguments and return bool.
 typedef BoolCallback = bool Function();
+
+/// There are two different dialog styles: Cupertino and Material
+enum UpgradeDialogStyle { cupertino, material }
 
 /// A class to define the configuration for the appcast. The configuration
 /// contains two parts: a URL to the appcast, and a list of supported OS
@@ -91,6 +95,9 @@ class Upgrader {
   /// The minimum app version supported by this app. Earlier versions of this app
   /// will be forced to update to the current version. Optional.
   String minAppVersion;
+
+  /// The upgrade dialog style. Optional. Used only on UpgradeAlert. (default: material)
+  UpgradeDialogStyle dialogStyle = UpgradeDialogStyle.material;
 
   bool _displayed = false;
   bool _initCalled = false;
@@ -248,17 +255,21 @@ class Upgrader {
 
   String message() {
     var msg = messages.message(UpgraderMessage.body);
-    msg = msg.replaceAll('{{appName}}', appName());
-    msg =
-        msg.replaceAll('{{currentAppStoreVersion}}', currentAppStoreVersion());
+    msg = msg.replaceAll('{{appName}}', appName() ?? '');
     msg = msg.replaceAll(
-        '{{currentInstalledVersion}}', currentInstalledVersion());
+        '{{currentAppStoreVersion}}', currentAppStoreVersion() ?? '');
+    msg = msg.replaceAll(
+        '{{currentInstalledVersion}}', currentInstalledVersion() ?? '');
     return msg;
   }
 
   void checkVersion({@required BuildContext context}) {
     if (!_displayed) {
-      if (shouldDisplayUpgrade()) {
+      final shouldDisplay = shouldDisplayUpgrade();
+      if (debugLogging) {
+        print('upgrader: shouldDisplayUpgrade: $shouldDisplay');
+      }
+      if (shouldDisplay) {
         _displayed = true;
         Future.delayed(Duration(milliseconds: 0), () {
           _showDialog(
@@ -277,7 +288,15 @@ class Upgrader {
 
   bool shouldDisplayUpgrade() {
     final isBlocked = blocked();
-    // If installed version below minimum app version, or is a critical update,
+
+    if (debugLogging) {
+      print('upgrader: blocked: $isBlocked');
+      print('upgrader: debugDisplayAlways: $debugDisplayAlways');
+      print('upgrader: debugDisplayOnce: $debugDisplayOnce');
+      print('upgrader: hasAlerted: $_hasAlerted');
+    }
+
+    // If installed version is below minimum app version, or is a critical update,
     // disable ignore and later buttons.
     if (isBlocked) {
       showIgnore = false;
@@ -342,6 +361,7 @@ class Upgrader {
       if (debugLogging) {
         print('upgrader: appStoreVersion: $_appStoreVersion');
         print('upgrader: installedVersion: $_installedVersion');
+        print('upgrader: minAppVersion: $minAppVersion');
         print('upgrader: isUpdateAvailable: $available');
       }
     }
@@ -385,35 +405,69 @@ class Upgrader {
       barrierDismissible: canDismissDialog,
       context: context,
       builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              Text(message),
-              Padding(
-                  padding: EdgeInsets.only(top: 15.0),
-                  child: Text(messages.message(UpgraderMessage.prompt))),
-            ],
-          ),
-          actions: <Widget>[
-            if (showIgnore)
-              FlatButton(
-                  child:
-                      Text(messages.message(UpgraderMessage.buttonTitleIgnore)),
-                  onPressed: () => onUserIgnored(context, true)),
-            if (showLater)
-              FlatButton(
-                  child:
-                      Text(messages.message(UpgraderMessage.buttonTitleLater)),
-                  onPressed: () => onUserLater(context, true)),
-            FlatButton(
-                child:
-                    Text(messages.message(UpgraderMessage.buttonTitleUpdate)),
-                onPressed: () => onUserUpdated(context, !blocked())),
-          ],
-        );
+        return dialogStyle == UpgradeDialogStyle.material
+            ? _alertDialog(title, message, context)
+            : _cupertinoAlertDialog(title, message, context);
       },
+    );
+  }
+
+  AlertDialog _alertDialog(String title, String message, BuildContext context) {
+    return AlertDialog(
+      title: Text(title),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          Text(message),
+          Padding(
+              padding: EdgeInsets.only(top: 15.0),
+              child: Text(messages.message(UpgraderMessage.prompt))),
+        ],
+      ),
+      actions: <Widget>[
+        if (showIgnore)
+          FlatButton(
+              child: Text(messages.message(UpgraderMessage.buttonTitleIgnore)),
+              onPressed: () => onUserIgnored(context, true)),
+        if (showLater)
+          FlatButton(
+              child: Text(messages.message(UpgraderMessage.buttonTitleLater)),
+              onPressed: () => onUserLater(context, true)),
+        FlatButton(
+            child: Text(messages.message(UpgraderMessage.buttonTitleUpdate)),
+            onPressed: () => onUserUpdated(context, !blocked())),
+      ],
+    );
+  }
+
+  CupertinoAlertDialog _cupertinoAlertDialog(
+      String title, String message, BuildContext context) {
+    return CupertinoAlertDialog(
+      title: Text(title),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: <Widget>[
+          Text(message),
+          Padding(
+              padding: EdgeInsets.only(top: 15.0),
+              child: Text(messages.message(UpgraderMessage.prompt))),
+        ],
+      ),
+      actions: <Widget>[
+        if (showIgnore)
+          CupertinoDialogAction(
+              child: Text(messages.message(UpgraderMessage.buttonTitleIgnore)),
+              onPressed: () => onUserIgnored(context, true)),
+        if (showLater)
+          CupertinoDialogAction(
+              child: Text(messages.message(UpgraderMessage.buttonTitleLater)),
+              onPressed: () => onUserLater(context, true)),
+        CupertinoDialogAction(
+            isDefaultAction: true,
+            child: Text(messages.message(UpgraderMessage.buttonTitleUpdate)),
+            onPressed: () => onUserUpdated(context, !blocked())),
+      ],
     );
   }