+++ test/upgrader_test.dart
@@ -540,9 +540,16 @@ void main() {
     verifyMessages(UpgraderMessages(code: 'ar'), 'ar');
     verifyMessages(UpgraderMessages(code: 'es'), 'es');
     verifyMessages(UpgraderMessages(code: 'fr'), 'fr');
+    verifyMessages(UpgraderMessages(code: 'de'), 'de');
+    verifyMessages(UpgraderMessages(code: 'hu'), 'hu');
+    verifyMessages(UpgraderMessages(code: 'id'), 'id');
+    verifyMessages(UpgraderMessages(code: 'it'), 'it');
     verifyMessages(UpgraderMessages(code: 'ko'), 'ko');
     verifyMessages(UpgraderMessages(code: 'pt'), 'pt');
     verifyMessages(UpgraderMessages(code: 'pl'), 'pl');
+    verifyMessages(UpgraderMessages(code: 'ru'), 'ru');
+    verifyMessages(UpgraderMessages(code: 'tr'), 'tr');
+    verifyMessages(UpgraderMessages(code: 'vi'), 'vi');
   });
 }